Extract v2 eval reference helpers
